+++ build/resourced/beforeremove.sh
@@ -1,15 +1,15 @@
 #!/bin/sh
 
-service cracklord-resourced stop >/dev/null 2>&1
+service cracklord-resourced stop >/dev/null 2>&1 || true
 
-if [ $1 = "remove" ]; then
-	if getent passwd cracklord >/dev/null ; then
-		userdel cracklord
-	fi
-
-	if getent group cracklord >/dev/null ; then
-		groupdel cracklord
-	fi
+if getent passwd cracklord >/dev/null ; then
+	userdel cracklord
 fi
 
-rm -r /var/cracklord
+if getent group cracklord >/dev/null ; then
+	groupdel cracklord
+fi
+
+if [ -d /var/cracklord ]; then
+	rm -r /var/cracklord
+fi
